# [Performance] Switch scheduled hot zone check to search instead of split