--- a/zone/zone_event_scheduler.cpp
+++ b/zone/zone_event_scheduler.cpp
@@ -34,11 +34,8 @@ void ZoneEventScheduler::Process(Zone *zone, WorldContentService *content_servic
 				LogSchedulerDetail("Looping active event validated [{}]", e.event_type);
 				if (e.event_type == ServerEvents::EVENT_TYPE_HOT_ZONE_ACTIVE) {
 					LogScheduler("Deactivating event [{}] disabling hotzone status", e.description);
-					for (auto &short_name: split(e.event_data, ',')) {
-						if (zone->GetShortName() == short_name) {
-							zone->SetIsHotzone(false);
-							break;
-						}
+					if (search_deliminated_string(e.event_data, zone->GetShortName()) != std::string::npos) {
+						zone->SetIsHotzone(false);
 					}
 					RemoveActiveEvent(e);
 				}
@@ -87,12 +84,9 @@ void ZoneEventScheduler::Process(Zone *zone, WorldContentService *content_servic
 			// such as broadcasts, reloads
 			if (ValidateEventReadyToActivate(e) && !IsEventActive(e)) {
 				if (e.event_type == ServerEvents::EVENT_TYPE_HOT_ZONE_ACTIVE) {
-					for (auto &short_name: split(e.event_data, ',')) {
-						if (zone->GetShortName() == short_name) {
-							zone->SetIsHotzone(true);
-							LogScheduler("Activating Event [{}] Enabling zone as hotzone", e.description);
-							break;
-						}
+					if (search_deliminated_string(e.event_data, zone->GetShortName()) != std::string::npos) {
+						zone->SetIsHotzone(true);
+						LogScheduler("Activating Event [{}] Enabling zone as hotzone", e.description);
 					}
 					m_active_events.push_back(e);
 				}