@Akkadius 2021-04-07
This looks great! This is a solid addition to our string utilities. I whipped together a quick test command myself

Keep in mind though that for the scope of the scheduler, this code get hits once a minute and is already in the nano-second realm

```cpp
	void TestCommand(int argc, char **argv, argh::parser &cmd, std::string &description)
	{
		description = "Test command";

		if (cmd[{"-h", "--help"}]) {
			return;
		}


		std::string event_data = "befallen,charasis,dalnir,frontiermtns,gukbottom,iceclad,lakeofillomen,northkarana,qey2hh1,soldunga,southro,wakening,podisease,velketor,akheva,riwwi,bothunder,poair";
		for (auto &search: split(event_data, ',')) {

			BenchTimer split_timer;

			bool      activated_split = false;
			for (auto &short_name: split(event_data, ',')) {
				if (search == short_name) {
					activated_split = true;
					break;
				}
			}

			BenchTimer search_timer;

			bool activated_deliminated = false;
			if (search_deliminated_string(event_data, search) != std::string::npos) {
				activated_deliminated = true;
			}

			LogInfo(
				"Searching for [{}] split ({}) took ({}s) search ({}) took ({}s)",
				search,
				(activated_split ? "true" : "false"),
				std::to_string(split_timer.elapsed()),
				(activated_split ? "true" : "false"),
				std::to_string(search_timer.elapsed())
			);
		}


	}
```

```
eqemu@e5311a8e9505:~/server$ ./bin/world test:test
[World] [Info] Searching for [befallen] split (true) took (0.000013s) search (true) took (0.000003s)
[World] [Info] Searching for [charasis] split (true) took (0.000003s) search (true) took (0.000000s)
[World] [Info] Searching for [dalnir] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [frontiermtns] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [gukbottom] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [iceclad] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [lakeofillomen] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [northkarana] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [qey2hh1] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [soldunga] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [southro] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [wakening] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [podisease] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [velketor] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [akheva] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [riwwi] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [bothunder] split (true) took (0.000002s) search (true) took (0.000000s)
[World] [Info] Searching for [poair] split (true) took (0.000002s) search (true) took (0.000000s)
```